auth/ui when the user sign out there is no message will display the user signed out there is no sign out form so query is changed. Instead of form the sign out message is aded
game/ui.
in get all games success function there was number of games players win
game status. that is deleted because that wont be displayed

app.js
whose turn message will not be displayed so it is deleted from document
ready function
and previously I forget console.log and it is deleted